pyproject.toml
[project]
name = "testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "imblearn>=0.0",
    "optuna>=4.7.0",
    "pandas>=3.0.1",
    "psycopg2>=2.9.11",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.47",
    "xgboost>=3.2.0",
]
